service/requirements.txt
Flask=0.11
Office365-REST-Python-Client
